# swagger.yml
swagger: "2.0"
info:
  title: yonder-api
  description: Official Yonder REST API
  version: 0.0.1
host: localhost:4000
basePath: /api
...
schemes:
  - http
  - https
paths:
...
  "/pay/methods":
    get:
      tags:
        - payment
      summary: Gets all payment methods for the authenticated user
      produces:
        - application/json
      responses:
        200:
          description: OK
        500:
          description: All other errors
    post:
      tags:
        - payment
      summary: Creates a payment method
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: Created
        409:
          description: Non-unique email
        500:
          description: NULL data, all other errors (at the moment)
    delete:
      tags:
        - payment
      summary: Deletes a payment method by id from the authenticated user
      produces:
        - application/json
      responses:
        200:
          description: Deleted
        500:
          description: All other errors

    "/pay/methods":
    get:
      tags:
        - payment
      summary: Gets all payment methods for the authenticated user
      produces:
        - application/json
      responses:
        200:
          description: OK
        500:
          description: All other errors
    post:
      tags:
        - payment
      summary: Creates a payment method
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: Created
        409:
          description: Non-unique email
        500:
          description: NULL data, all other errors (at the moment)
    delete:
      tags:
        - payment
      summary: Deletes a payment method by id from the authenticated user
      produces:
        - application/json
      responses:
        200:
          description: Deleted
        500:
          description: All other errors
...